fix: merge element-derived snap positions into the rendered slide
Wire the existing RenderedElement.snap_positions channel through the slide renderer into SlideHTML, so image-sequence frame snaps (already reported by 'scrolly introspect snaps' as the merged set) actually reach the built deck. Widen the rendered snap list to float. Previously an image-sequence slide with no explicit snap_positions snapped only to its endpoints.

+++ scrolly/slide/element_ir/rendered.py
@@ -33,6 +33,6 @@ class RenderedElement:
 
     html: str
     scoped_css: str = ""
-    snap_positions: tuple[int, ...] = ()
+    snap_positions: tuple[float, ...] = ()
     assets: tuple[Path, ...] = ()
     has_mermaid: bool = False
